+++ src/main/java/katecam/luvicookie/ditto/domain/login/config/Config.java
@@ -2,7 +2,7 @@ package katecam.luvicookie.ditto.domain.login.config;
 
 
 import katecam.luvicookie.ditto.domain.login.annotation.LoginUserArgumentResolver;
-import katecam.luvicookie.ditto.domain.member.service.MemberService;
+import katecam.luvicookie.ditto.domain.member.application.MemberService;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
