+++ src/main/java/katecam/luvicookie/ditto/domain/login/oauth/OAuth2UserCustomService.java
@@ -4,7 +4,7 @@ import katecam.luvicookie.ditto.domain.member.domain.KakaoUserInfo;
 import katecam.luvicookie.ditto.domain.member.domain.Member;
 import katecam.luvicookie.ditto.domain.member.domain.PrincipalDetail;
 import katecam.luvicookie.ditto.domain.member.domain.Role;
-import katecam.luvicookie.ditto.domain.member.repository.MemberRepository;
+import katecam.luvicookie.ditto.domain.member.dao.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
