+++ src/main/java/katecam/luvicookie/ditto/domain/member/api/MemberController.java
@@ -1,17 +1,14 @@
-package katecam.luvicookie.ditto.domain.member.controller;
+package katecam.luvicookie.ditto.domain.member.api;
 
-import katecam.luvicookie.ditto.domain.login.UrlResponse;
 import katecam.luvicookie.ditto.domain.login.annotation.LoginUser;
 import katecam.luvicookie.ditto.domain.member.domain.Member;
-import katecam.luvicookie.ditto.domain.member.domain.PrincipalDetail;
 import katecam.luvicookie.ditto.domain.member.dto.memberRequestDTO;
 import katecam.luvicookie.ditto.domain.member.dto.memberResponseDTO;
 import katecam.luvicookie.ditto.domain.member.dto.profileImageDTO;
-import katecam.luvicookie.ditto.domain.member.service.MemberService;
+import katecam.luvicookie.ditto.domain.member.application.MemberService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
