+++ src/main/java/katecam/luvicookie/ditto/domain/member/application/PrincipalDetailsService.java
@@ -1,8 +1,8 @@
-package katecam.luvicookie.ditto.domain.member.service;
+package katecam.luvicookie.ditto.domain.member.application;
 
 import katecam.luvicookie.ditto.domain.member.domain.Member;
 import katecam.luvicookie.ditto.domain.member.domain.PrincipalDetail;
-import katecam.luvicookie.ditto.domain.member.repository.MemberRepository;
+import katecam.luvicookie.ditto.domain.member.dao.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.userdetails.UserDetails;
